Extract sidebar workspace row
